refactor: Api 함수 파일명 통일
