+++ src/app/viewing/reservation/[id]/confirm/page.tsx
@@ -7,7 +7,7 @@ import { useState } from "react";
 import { useViewingScheduleStore } from "@/stores/useViewingScheduleStore";
 import { useQueryClient } from "@tanstack/react-query";
 
-import { createViewing } from "@/apis/viewing";
+import { createViewing } from "@/apis/viewingApi";
 
 import { usePropertyList } from "@/hooks/property/useProperty";
