+++ src/components/viewings/ViewingNoteSection.tsx
@@ -1,7 +1,7 @@
 import { forwardRef, useEffect, useImperativeHandle, useState } from "react";
 
-import { getPropertyNotePresignedUrl } from "@/apis/s3Upload";
-import { putViewingPropertyNotes } from "@/apis/viewing";
+import { getPropertyNotePresignedUrl } from "@/apis/s3UploadApi";
+import { putViewingPropertyNotes } from "@/apis/viewingApi";
 
 import { useViewingDetail } from "@/hooks/viewing/useViewing";
